Rename & improve blocks utility
