+++ src/utils/index.js
@@ -1,5 +1,5 @@
 import bech32Manager from './bech32Manager';
-import blocksRequestHandler from './blocksRequestHandler';
+import blocksRequestHelper from './blocksRequestHelper';
 import coinAdapter from './coinAdapter';
 import dateHandler from './dateHandler';
 import proposerHandler from './proposerHandler';
@@ -10,7 +10,7 @@ import txsTableAdapter from './txs/txsTableAdapter';
 
 export {
   bech32Manager,
-  blocksRequestHandler,
+  blocksRequestHelper,
   coinAdapter,
   dateHandler,
   proposerHandler,
